fix the loops and timers; use go 1.23
Requires Go 1.23 for timers reimplementation.

Previously, non-write/create events were restarting the command;
there was lot of them on darwin.
Plus there was series of write events creating restarts one by one.
This is fixed now.
Plus, current dir is added to the watcher with filtering events on the cmd file
so the file creation evenrs can be actually caught.